feat: webpack for local development
