+++ app/index.html
@@ -5,7 +5,7 @@
     <title>Voice of InterConnect</title>
   </head>
   <body>
-    <h1>Voice of InterConnect</h1>
+    <h1>Voice of InterConnect!</h1>
 
     <button id="record">Simulate recording note</button>
 
@@ -16,8 +16,5 @@
     <p id="status">Loading ...</p>
 
     <table id="notes"></table>
-
-    <script src="/hoodie/client.js"></script>
-    <script src="/assets/app.js"></script>
   </body>
 </html>
