+++ app/lib/set-app-status.js
@@ -1,0 +1,5 @@
+module.exports = setAppStatus
+
+function setAppStatus (status) {
+  document.querySelector('#status').textContent = status
+}
